+++ DarmuhsTerminalCommands/VisualCore/SplitViewChecks.cs
@@ -1,87 +1,111 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.UI;
+using UnityEngine.Video;
 using static TerminalStuff.EventSub.TerminalQuit;
+using static TerminalStuff.BoolStuff;
+using OpenLib.CoreMethods;
 
 namespace TerminalStuff
 {
     internal class SplitViewChecks : MonoBehaviour
     {
         internal static bool enabledSplitObjects = false;
+        internal static CompatibleNoun[] vanillaViewNouns;
+        internal static TerminalKeyword viewKeyword;
+
+        private static void HandleVanillaMap(bool shouldRemove)
+        {
+            if(shouldRemove)
+            {
+                if (DynamicBools.TryGetKeyword("view", out TerminalKeyword viewKW) && DynamicBools.TryGetKeyword("monitor", out TerminalKeyword monitorKW))
+                {
+                    Plugin.Spam("removing vanilla compat noun - monitor, from view");
+                    viewKeyword = viewKW;
+                    vanillaViewNouns = viewKW.compatibleNouns;
+                    RemoveThings.RemoveCompatibleNoun(ref viewKW, monitorKW);
+                }
+            }
+            else
+            {
+                if(vanillaViewNouns != null)
+                {
+                    Plugin.Spam("returning to vanilla compat nouns");
+                    viewKeyword.compatibleNouns = vanillaViewNouns;
+                }
+            }
+        }
 
         public static void InitSplitViewObjects()
         {
-            Plugin.instance.terminalCanvas = Plugin.instance.Terminal.terminalUIScreen;
-
-            if (Plugin.instance.terminalCanvas == null)
+            if (!ShouldAddCamsLogic())
             {
-                Plugin.Log.LogError("Canvas not found");
+                HandleVanillaMap(false);
+                return;
+            }
+            else
+                HandleVanillaMap(true);
+                
+
+            if (Plugin.instance.Terminal.terminalImage == null || Plugin.instance.splitViewCreated)
+            {
+                Plugin.MoreLogs("Original terminalImage not found or split view already created");
                 return;
             }
 
-            Plugin.MoreLogs("Canvas found");
+            //ViewCommands.miniScreen = Instantiate(Plugin.instance.Terminal.terminalImage.gameObject, Plugin.instance.Terminal.terminalImage.gameObject.transform);
+            //ViewCommands.miniScreen.name = "Terminal Small Screen (Clone)";
+            //ViewCommands.miniScreenImage = ViewCommands.miniScreen.GetComponent<RawImage>();
 
-            RawImage originalRawImage = Plugin.instance.terminalCanvas.transform.Find("MainContainer/ImageContainer/Image (1)")?.GetComponent<RawImage>();
-
-            if (originalRawImage == null || Plugin.instance.splitViewCreated)
+            ViewCommands.miniScreenImage = Instantiate(Plugin.instance.Terminal.terminalImage, Plugin.instance.Terminal.terminalImage.transform);
+            
+            if(ViewCommands.miniScreenImage.gameObject.GetComponent<VideoPlayer>() != null)
             {
-                Plugin.MoreLogs("Original RawImage not found or split view already created");
-                return;
+                VideoPlayer extraPlayer = ViewCommands.miniScreenImage.GetComponent<VideoPlayer>();
+                Destroy(extraPlayer);
+                Plugin.Spam("extraPlayer deleted");
             }
 
-            Plugin.MoreLogs("Original RawImage found");
-
-            // Duplicate the RawImage
-            GameObject rawImageGameObject2 = Instantiate(originalRawImage.gameObject, originalRawImage.transform.parent);
-            GameObject rawImageGameObject1 = Instantiate(originalRawImage.gameObject, originalRawImage.transform.parent);
-            rawImageGameObject1.name = "Terminal Small Screen (Clone)";
-            rawImageGameObject2.name = "Terminal Full Screen (Clone)";
-
-            // Get the RawImage components from the duplicated GameObjects
-            Plugin.instance.rawImage2 = rawImageGameObject2.GetComponent<RawImage>();
-            Plugin.instance.rawImage1 = rawImageGameObject1.GetComponent<RawImage>();
-
-            // Store the original dimensions and anchored positions
-            StoreOriginalDimensionsAndPositions();
+            ViewCommands.miniScreenImage.gameObject.name = "MiniScreen";
+            Plugin.instance.Terminal.terminalImage.gameObject.name = "terminalImage";
 
             Plugin.instance.splitViewCreated = true;
         }
 
-        private static void StoreOriginalDimensionsAndPositions()
-        {
-            Plugin.instance.originalTopSize = Plugin.instance.rawImage1.rectTransform.sizeDelta;
-            Plugin.instance.originalTopPosition = Plugin.instance.rawImage1.rectTransform.anchoredPosition;
-
-            Plugin.instance.originalBottomSize = Plugin.instance.rawImage2.rectTransform.sizeDelta;
-            Plugin.instance.originalBottomPosition = Plugin.instance.rawImage2.rectTransform.anchoredPosition;
-        }
-
         public static void CheckForSplitView(string whatIsIt)
         {
-            // Assuming you have a reference to the ImageContainer GameObject
-            GameObject imageContainer = Plugin.instance.Terminal.terminalImageMask.gameObject;
+            if (!Plugin.instance.splitViewCreated && whatIsIt == "neither")
+            {
+                DisableVanillaViewMonitor();
+                return;
+            }
+            else if (!Plugin.instance.splitViewCreated)
+                return;
+                
             List<string> singleViewModes = ["cams", "map", "mirror"];
 
             if (enabledSplitObjects == false)
             {
-                DisableCloneChildren(imageContainer);
+                SetMiniScreen(whatIsIt, false);
                 ResetPluginInstanceBools();
             }
             else if (whatIsIt == "neither")
             {
                 enabledSplitObjects = false;
-                DisableCloneChildren(imageContainer);
+                //DisableVanillaViewMonitor();
+                SetMiniScreen(whatIsIt, false);
                 ResetPluginInstanceBools();
             }
             else if (enabledSplitObjects == true && (!singleViewModes.Contains(whatIsIt)))
             {
-                EnableCloneChildren(imageContainer);
+                //DisableVanillaViewMonitor(disableImage: false);
+                SetMiniScreen(whatIsIt, true);
                 ViewCommands.SetAnyCamsTrue();
                 UpdatePluginInstanceBools(whatIsIt);
             }
             else if (enabledSplitObjects == true && (singleViewModes.Contains(whatIsIt)))
             {
-                EnableFullScreenAndDisableSmallScreen(imageContainer, whatIsIt);
+                //DisableVanillaViewMonitor(disableImage: false);
+                SetMiniScreen(whatIsIt, false);
                 UpdatePluginInstanceBools(whatIsIt);
             }
             else
@@ -90,52 +114,12 @@ namespace TerminalStuff
             }
         }
 
-        private static void DisableCloneChildren(GameObject imageContainer)
+        private static void SetMiniScreen(string whatIsIt, bool state)
         {
-            if (imageContainer != null)
-            {
-                foreach (Transform child in imageContainer.transform)
-                {
-                    if (child.gameObject.name.Contains("Terminal Full Screen (Clone)") || child.gameObject.name.Contains("Terminal Small Screen (Clone)"))
-                    {
-                        child.gameObject.SetActive(false);
-                    }
-                }
-            }
-        }
-
-        private static void EnableCloneChildren(GameObject imageContainer)
-        {
-            if (imageContainer != null)
-            {
-                foreach (Transform child in imageContainer.transform)
-                {
-                    if (child.gameObject.name.Contains("Terminal Full Screen (Clone)") || child.gameObject.name.Contains("Terminal Small Screen (Clone)"))
-                    {
-                        child.gameObject.SetActive(true);
-                    }
-                }
-            }
-        }
-
-        private static void EnableFullScreenAndDisableSmallScreen(GameObject imageContainer, string whatIsIt)
-        {
-            if (imageContainer != null)
-            {
-                foreach (Transform child in imageContainer.transform)
-                {
-                    if (child.gameObject.name.Equals("Terminal Full Screen (Clone)"))
-                    {
-                        child.gameObject.SetActive(true);
-                        Plugin.MoreLogs($"Enabled full screen for {whatIsIt}");
-                    }
-                    if (child.gameObject.name.Equals("Terminal Small Screen (Clone)"))
-                    {
-                        child.gameObject.SetActive(false);
-                        Plugin.MoreLogs($"Disabled small screen for {whatIsIt}");
-                    }
-                }
-            }
+            //Plugin.instance.Terminal.terminalImage.enabled = state;
+            ViewCommands.miniScreenImage.enabled = state;
+            Plugin.MoreLogs($"{whatIsIt} has set miniscreen to {state}");
+            //ViewCommands.miniScreen.gameObject.SetActive(state);
         }
 
         internal static void ResetPluginInstanceBools()
@@ -151,8 +135,24 @@ namespace TerminalStuff
             ShowCameraView(true);
         }
 
+        internal static void DisableVanillaViewMonitor(bool disableImage = true)
+        {
+            Plugin.Spam("disabling vanilla view monitor");
+            Plugin.instance.Terminal.displayingPersistentImage = null;
+            if(disableImage) 
+                Plugin.instance.Terminal.terminalImage.enabled = false;
+        }
+
         internal static void DisableSplitView(string whatIsIt)
         {
+            if (!Plugin.instance.splitViewCreated && whatIsIt == "neither")
+            {
+                DisableVanillaViewMonitor();
+                return;
+            }
+            else if (!Plugin.instance.splitViewCreated)
+                return;
+
             enabledSplitObjects = false;
             CheckForSplitView(whatIsIt);
             ShowCameraView(false);